Move "RestClient" in sub folder
